requirements.txt
mysql-python
django-annoying
pillow

django_extensions
Werkzeug
django-debug_toolbar
